Fixing the test functions
The test now run flawlessly.

The problem with the checks of the server availability was a bot
filter running on the server of the DWD. By waiting a small fraction
of a second before launching the next request, this detection is not
triggered and the test works again. I tried several times, including
stochastic ones, and 0.00001 seconds did work best.

In addition, I fixed the clean up procedure. In there a global option
of R was altered but not changed back to its original state.