Fixing download of the content of the FTP server
Occasionally **wget** is not able to download the files from the FTP
server in the **download.data.dwd** function. It does not return the
same error message every time but restarting the download usually
fixes the issue.

In the current implementation a *try-catch* statement is introduced in
addition with a constant specifying the maximum number of tries. If
the download isn't working, the script will try it again 9 times and
if it still does not work, the file is skipped.