Restructuring the extraction and import
The **download.data.dwd** function was moved to the *data-download.R*
file and the extraction and conversion part was separated and put
into the **conversion.climate** function
in *import.R*. The **extract.content** function was renamed
to **extract.content.climate** and was along
with **extract.station.names** moved into separate functions.

The extraction of the recent and historical files in separate folders
does already work. But apart from this it is still a big mess.