Fix extraction and split results into climatological objects
A small bug prevented the **extract.content.climate** function to find
the unique archives corresponding to a station id.

In addition, the list of stations, where each station is itself a list
containing the climatological quantities as elements, was converted
into lists of climatological elements. Those are split by name into
different objects. Their names are prefixed with *dwd.*

Descriptive names for the aggregated data were introduced in
the **extract.content.climate** function.